refactor: centralize build path resolution
