+++ autoresearch/eval-save.ts
@@ -80,7 +80,7 @@ function judge(criteria: JudgeCriteria, output: string): boolean {
 
 const PROJECT_ROOT = join(__dirname, '..');
 
-/** Run a command, using local dist/main.js instead of global opencli for consistency */
+/** Run a command, using the local built entrypoint instead of global opencli for consistency */
 function runCommand(cmd: string, timeout = 30000): string {
   // Use local build so tests always run against the current source
   const localCmd = cmd.replace(/^opencli /, `node dist/src/main.js `);
