+++ src/cli.ts
@@ -10,6 +10,7 @@ import * as path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { Command } from 'commander';
 import chalk from 'chalk';
+import { findPackageRoot, getBuiltEntryCandidates } from './package-paths.js';
 import { type CliCommand, fullName, getRegistry, strategyLabel } from './registry.js';
 import { serializeCommand, formatArgSummary } from './serialization.js';
 import { render as renderOutput } from './output.js';
@@ -1016,48 +1017,7 @@ export interface OperateVerifyInvocation {
   shell?: boolean;
 }
 
-export function findPackageRoot(startFile: string, fileExists: (path: string) => boolean = fs.existsSync): string {
-  let dir = path.dirname(startFile);
-
-  while (true) {
-    if (fileExists(path.join(dir, 'package.json'))) return dir;
-    const parent = path.dirname(dir);
-    if (parent === dir) {
-      throw new Error(`Could not find package.json above ${startFile}`);
-    }
-    dir = parent;
-  }
-}
-
-function getBuiltEntryCandidates(packageRoot: string, readFile: (path: string) => string): string[] {
-  const candidates: string[] = [];
-  try {
-    const pkg = JSON.parse(readFile(path.join(packageRoot, 'package.json'))) as {
-      bin?: string | Record<string, string>;
-      main?: string;
-    };
-
-    if (typeof pkg.bin === 'string') {
-      candidates.push(path.join(packageRoot, pkg.bin));
-    } else if (pkg.bin && typeof pkg.bin === 'object' && typeof pkg.bin.opencli === 'string') {
-      candidates.push(path.join(packageRoot, pkg.bin.opencli));
-    }
-
-    if (typeof pkg.main === 'string') {
-      candidates.push(path.join(packageRoot, pkg.main));
-    }
-  } catch {
-    // Fall through to compatibility candidates below.
-  }
-
-  // Compatibility fallback for partially-built trees or older layouts.
-  candidates.push(
-    path.join(packageRoot, 'dist', 'src', 'main.js'),
-    path.join(packageRoot, 'dist', 'main.js'),
-  );
-
-  return [...new Set(candidates)];
-}
+export { findPackageRoot };
 
 export function resolveOperateVerifyInvocation(opts: {
   projectRoot?: string;
